+++ src/Billing/BillingInterface.php
@@ -2,9 +2,7 @@
 
 namespace Hyvor\Internal\Billing;
 
-use Hyvor\Internal\Billing\Dto\LicenseOf;
-use Hyvor\Internal\Billing\Dto\LicensesCollection;
-use Hyvor\Internal\Billing\License\License;
+use Hyvor\Internal\Billing\License\Resolved\ResolvedLicense;
 use Hyvor\Internal\Component\Component;
 use Hyvor\Internal\InternalApi\Exceptions\InternalApiCallFailedException;
 
@@ -13,11 +11,12 @@ interface BillingInterface
     /**
      * @throws InternalApiCallFailedException
      */
-    public function license(int $organizationId, ?int $resourceId, ?Component $component = null): ?License;
+    public function license(int $organizationId, ?Component $component = null): ResolvedLicense;
 
     /**
-     * @param array<LicenseOf> $of
+     * @param int[] $organizationIds
+     * @return array<int, ResolvedLicense> orgId keyed licenses
      * @throws InternalApiCallFailedException
      */
-    public function licenses(array $of, ?Component $component = null): LicensesCollection;
+    public function licenses(array $organizationIds, ?Component $component = null): array;
 }