+++ src/Auth/AuthCurrentOrganization.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Hyvor\Internal\Auth;
+
+class AuthCurrentOrganization
+{
+
+    public function __construct(
+        public int $id,
+        public string $name,
+    ) {
+    }
+
+}