+++ tests/Bundle/Api/AbstractApiExceptionListenerTest.php
@@ -2,7 +2,7 @@
 
 namespace Hyvor\Internal\Tests\Bundle\Api;
 
-use Hyvor\Internal\Bundle\Api\AbstractApiExceptionListener;
+use Hyvor\Internal\Bundle\Api\ApiExceptionListener;
 use Hyvor\Internal\Bundle\Api\DataCarryingHttpException;
 use Hyvor\Internal\Tests\SymfonyTestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -15,7 +15,7 @@ use Symfony\Component\Validator\ConstraintViolation;
 use Symfony\Component\Validator\ConstraintViolationList;
 use Symfony\Component\Validator\Exception\ValidationFailedException;
 
-#[CoversClass(AbstractApiExceptionListener::class)]
+#[CoversClass(ApiExceptionListener::class)]
 #[CoversClass(DataCarryingHttpException::class)]
 class AbstractApiExceptionListenerTest extends SymfonyTestCase
 {
@@ -195,25 +195,4 @@ class AbstractApiExceptionListenerTest extends SymfonyTestCase
         $this->assertSame(['foo' => 'bar'], $data['data']);
     }
 
-}
-
-
-class ApiExceptionListener extends AbstractApiExceptionListener
-{
-    protected function prefix(): string
-    {
-        return '/api/console';
-    }
-}
-
-
-// Helpers
-
-namespace App\Enum;
-
-enum TestEnum: string
-{
-    case TEST = 'test';
-    case TEST2 = 'test2';
-    case TEST3 = 'test3';
 }