+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - BillingInterface license() and licenses() method signatures changed
 - Comms API
   - a shared COMMS_KEY env variable is required for component-to-componet communication
+- ResourceInterface and resource registering in the core is removed
 
 ## 3.1.8 - 2025-11-03