+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 4.0.0 - Unreleased
+
+- Organization support
+- Hyvor\Internal\Laravel\LogFake removed
+- TestEventDispatcher moved
+- BillingInterface license() and licenses() method signatures changed
+- Comms API
+  - a shared COMMS_KEY env variable is required for component-to-componet communication
+- ResourceInterface and resource registering in the core is removed
+
 ## 3.1.8 - 2025-11-03
 
 - Self-Hosted telemetry support added